refactor(base): update exports to work with Vite
