+++ packages/@sanity/base/src/client/index.esm.ts
@@ -1,0 +1,1 @@
+export {wrappedClient as default} from './wrappedClient'
